Fix SQLite errors Fix slow speed caused by ldap auth

+++ src/multi-chat/database/migrations/2024_02_04_162131_delete_some_columns_that_doesnot_need.php
@@ -14,13 +14,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('chats', function (Blueprint $table) {
-            $table->dropColumn('name');
+            if (Schema::hasColumn('chats', 'name')) {
+                $table->dropColumn('name');
+            }
         });
-        Schema::table('llms', function (Blueprint $table) {
-            $table->dropColumn('link');
-            $table->dropColumn('healthy');
-            $table->dropColumn('version');
-        });
+        
     }
 
     /**
